- criação de middleware para validação dos payloads - criação de utilitário para responder erros - ajuste nos testes
