+++ routes/pets.js
@@ -1,8 +1,8 @@
-const { authentication: AuthenticationMiddleware } = require('../middlewares')
+const { authentication: authenticationMiddleware, 'payload-validator': validatorMiddleware } = require('../middlewares')
 const { pets: petsController } = require('../controllers')
 
 module.exports = (app) => {
-  app.post('/pets', AuthenticationMiddleware.validateAuthentication, petsController.create)
-  app.get('/pets', AuthenticationMiddleware.validateAuthentication, petsController.search)
-  app.get('/pets/:id', AuthenticationMiddleware.validateAuthentication, petsController.find)
+  app.post('/pets', authenticationMiddleware.validateAuthentication, validatorMiddleware.validatePetPayload, petsController.create)
+  app.get('/pets', authenticationMiddleware.validateAuthentication, validatorMiddleware.validateSearchPayload, petsController.search)
+  app.get('/pets/:id', authenticationMiddleware.validateAuthentication, petsController.find)
 }
