+++ utils/respondError.js
@@ -1,0 +1,3 @@
+module.exports = (res, error) => {
+  return res.status(error.status).send({message: error.message})
+}
